Fix lint issues in steam notebook
