+++ .gitignore
@@ -1,1 +1,5 @@
 .ipynb_checkpoints
+.pytest_cache
+.ruff_cache
+
+__pycache__
