+++ 03 - toomics/main.py
@@ -21,9 +21,8 @@
 
 
 import requests
-from framework import get, get_soup
+from framework import get_soup
 import pandas as pd
-from bs4 import BeautifulSoup
 
 
 # ### **Unlimited Power**
