+++ 04 - jmlr/main.py
@@ -16,11 +16,9 @@
 
 import re
 import pandas as pd
-import requests
-from framework import get, get_soup
+from framework import get_soup
 
 from tqdm.notebook import tqdm
-from bs4 import BeautifulSoup
 
 
 # ### **Variables**
