+++ 10 - a24 [uncompleted]/main.py
@@ -46,10 +46,8 @@
 
 
 import json
-import requests
-from framework import get, get_soup, post
+from framework import get_soup
 
-from bs4 import BeautifulSoup
 from math import ceil
 
 
@@ -143,7 +141,6 @@ def get_title_information(url: str):
     if meta['productionStatus']['currentProductionStage']['id'] == 'in_development':
         return
 
-    from pprint import pprint
     #pprint(meta)
 
     # data['type'] = meta['@type']
